Various Fixes
Fixed a couple mistakes. We should not be trying to get the host url from requests as this is a cron process and not being triggered by a request and thus the "has_request_context" will never be true. Instead use our DOMAIN value that should be set during fireshare setup. 

Fixed an issue where the video_id was not being passed to the get_public_watch_url and instead a hardcoded testing id was.